+++ app/sprinkles/core/src/ServicesProvider/ServicesProvider.php
@@ -128,7 +128,7 @@ class ServicesProvider
                 $locator->registerStream('assets', 'vendor', \UserFrosting\BOWER_ASSET_DIR, true);
                 $locator->registerStream('assets', 'vendor', \UserFrosting\NPM_ASSET_DIR, true);
                 $locator->registerStream('assets', 'vendor', \UserFrosting\BROWSERIFIED_ASSET_DIR, true);
-                $locator->registerStream('assets', '', \UserFrosting\DS . \UserFrosting\ASSET_DIR_NAME);
+                $locator->registerStream('assets', '', \UserFrosting\ASSET_DIR_NAME);
 
                 $baseUrl = $config['site.uri.public'] . '/' . $config['assets.raw.path'];
 
@@ -698,8 +698,9 @@ class ServicesProvider
             $sprinkles = $c->sprinkleManager->getSprinkleNames();
 
             // Add Sprinkles' templates namespaces
+            // TODO : Use locator
             foreach ($sprinkles as $sprinkle) {
-                $path = \UserFrosting\SPRINKLES_DIR . \UserFrosting\DS .
+                $path = \UserFrosting\APP_DIR . \UserFrosting\DS . \UserFrosting\SPRINKLES_DIR_NAME . \UserFrosting\DS .
                     $sprinkle . \UserFrosting\DS .
                     \UserFrosting\TEMPLATE_DIR_NAME . \UserFrosting\DS;